requirements.txt
torch
torchvision
opencv-python
yolov5
onvif-zeep
matplotlib
pyrealsense2
# oafd oubq vziv aqub